refactor: move extra routes to routes mod
