+++ crates/service/src/routes/health.rs
