+++ crates/service/src/routes/mod.rs
@@ -3,6 +3,12 @@
 
 pub mod cost;
 pub mod dips;
+mod health;
+mod request_handler;
+mod static_subgraph;
 mod status;
 
+pub use health::health;
+pub use request_handler::request_handler;
+pub use static_subgraph::static_subgraph_request_handler;
 pub use status::status;
